requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pandas
openpyxl
requests
python-multipart
